# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Clarity B2B API",
    "version": "1.0.0",
    "description": "API documentation for the Clarity Supply Chain Finance backend. This covers the Invoice Pipeline and Supplier Health Analytics modules."
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/erp/ledger/{invoiceId}": {
      "patch": {
        "tags": ["ERP"],
        "summary": "Update a ledger row",
        "description": "Updates a manual payable or accounting metadata on a platform-synced row.",
        "parameters": [{ "name": "invoiceId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Ledger row updated" } }
      },
      "delete": {
        "tags": ["ERP"],
        "summary": "Delete manual payable",
        "description": "Deletes a manual payable; platform rows remain source-managed.",
        "parameters": [{ "name": "invoiceId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Manual payable deleted" } }
      }
    },
...
  }
}